backend/requirements.txt
fastapi
uvicorn
python-multipart
qrcode[pil]
pillow
numpy